requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
mssql-django==1.3
pycparser==2.21
PyJWT==2.4.0
pyodbc==5.0.0
pytz==2022.1
requests==2.31.0
sqlparse==0.4.2
typing-extensions==4.8.0
tzdata==2022.1
urllib3==2.0.6
boto3 == 1.28.63
rembg == 2.0.50
load_dotenv == 1.0.1
dotenv == 0.0.5
